+++ examples/nestjs-basic/src/user.repository.ts
@@ -1,9 +1,7 @@
 import { createInjectableRepository } from '@prismakit/nestjs';
 
-export const UserRepository = createInjectableRepository({
+export class UserRepository extends createInjectableRepository({
   model: 'user',
   scalarFields: { id: 'id', email: 'email', name: 'name' },
   cache: { ttl: 60 },
-});
-
-export type UserRepository = InstanceType<typeof UserRepository>;
+}) {}