+++ packages/benchmark/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prismakit/benchmark
 
+## 2.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @prismakit/core@3.0.2
+
 ## 2.2.3
 
 ### Patch Changes