+++ packages/memory/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prismakit/memory
 
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @prismakit/core@3.0.2
+
 ## 3.0.0
 
 ### Major Changes